Promote the subscribeBy extension functions from recipes to part of the Kotlin artifact.
